refactor: detail 페이지 폴더 생성후 분리 #82
