+++ frontend/src/pages/detail/components/DetailHeader/DetailHeader.stories.tsx
