+++ frontend/src/pages/detail/components/DetailHeader/DetailHeader.tsx
